pyproject.toml
[project]
name = "cloud_speed_test"
version = "0.0.1"
description = "Generic framework for running speed tests on cloud architecture"
authors = [
    { name = "Bob Clapp", email = "clapp@stanford.edu" }
]

dependencies=[ 'google-cloud-compute>=1.10.0',
                'pandas>=2.0',
            ]
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]
